+++ patientMatcher/cli/update.py
@@ -1,16 +1,59 @@
 #!/usr/bin/env python3
 # -*- coding: utf-8 -*-
 
+import click
 import requests
 from clint.textui import progress
-import click
+from flask.cli import current_app, with_appcontext
 from patientMatcher.constants import PHENOTYPE_TERMS
+from patientMatcher.utils.patient import patients
 
 
 @click.group()
 def update():
     """Update patientMatcher resources"""
     pass
+
+
+@update.command()
+@with_appcontext
+@click.option("-old-href", type=click.STRING, nargs=1, required=True, help="Old contact href")
+@click.option("-href", type=click.STRING, nargs=1, required=True, help="New contact href")
+@click.option("-name", type=click.STRING, nargs=1, required=True, help="New contact name")
+@click.option(
+    "-institution", type=click.STRING, nargs=1, required=False, help="New contact institution"
+)
+def contact(old_href, href, name, institution):
+    """Update contact person for a group of patients"""
+
+    database = current_app.db
+    query = {"contact.href": {"$regex": old_href}}
+
+    # Retrieving all patients matching the given old_href
+    old_contact_patients = patients(database=database, match_query=query)
+    # Retriving unique contacts for the above patients
+    match_contacts = list(old_contact_patients.distinct("contact.href"))
+
+    if len(match_contacts) == 0:
+        click.echo(f"No patients found with contact URI '{old_href}'")
+        return
+    if len(match_contacts) > 1:
+        click.echo(
+            f"Your search for contact url '{old_href}' is returning more than one patients' contact: {match_contacts}.\nPlease restrict your search by typing a different href."
+        )
+        return
+    # Search is returning only one contact, it's OK to use it for updating patients
+    matches = list(old_contact_patients)
+    new_contact = dict(href=href, name=name)
+    if institution:
+        new_contact["institution"] = institution
+
+    if click.confirm(
+        f"{len(matches)} patients with the old contact href '{matches[0]['contact']['href']}' will be updated with contact info:{new_contact}. Confirm?",
+        abort=True,
+    ):
+        result = database.patients.update_many(query, {"$set": {"contact": new_contact}})
+        click.echo(f"Contact information was updated for {result.modified_count} patients.")
 
 
 @update.command()